+++ src/widgets/button.rs
@@ -1,8 +1,8 @@
 //! A clickable, labeled button
 use std::time::{Duration, Instant};
 
-use figures::units::{Lp, Px, UPx};
-use figures::{IntoSigned, Point, Rect, Round, ScreenScale, Size};
+use figures::units::{Px, UPx};
+use figures::{IntoSigned, Point, Rect, Round, ScreenScale, Size, Zero};
 use kludgine::app::winit::event::{Modifiers, MouseButton};
 use kludgine::app::winit::window::CursorIcon;
 use kludgine::shapes::{Shape, StrokeOptions};
@@ -11,10 +11,11 @@ use kludgine::Color;
 use crate::animation::{AnimationHandle, AnimationTarget, IntoAnimate, LinearInterpolate, Spawn};
 use crate::context::{AsEventContext, EventContext, GraphicsContext, LayoutContext, WidgetContext};
 use crate::styles::components::{
-    AutoFocusableControls, DefaultActiveBackgroundColor, DefaultActiveForegroundColor,
-    DefaultBackgroundColor, DefaultDisabledBackgroundColor, DefaultDisabledForegroundColor,
-    DefaultForegroundColor, DefaultHoveredBackgroundColor, DefaultHoveredForegroundColor, Easing,
-    HighlightColor, IntrinsicPadding, OpaqueWidgetColor, OutlineColor, SurfaceColor, TextColor,
+    AutoFocusableControls, CornerRadius, DefaultActiveBackgroundColor,
+    DefaultActiveForegroundColor, DefaultBackgroundColor, DefaultDisabledBackgroundColor,
+    DefaultDisabledForegroundColor, DefaultForegroundColor, DefaultHoveredBackgroundColor,
+    DefaultHoveredForegroundColor, Easing, HighlightColor, IntrinsicPadding, OpaqueWidgetColor,
+    OutlineColor, OutlineWidth, SurfaceColor, TextColor,
 };
 use crate::styles::{ColorExt, Styles};
 use crate::value::{Destination, Dynamic, IntoValue, Source, Value};
@@ -370,12 +371,16 @@ impl Widget for Button {
         let style = self.current_style(context);
         context.fill(style.background);
 
-        let two_lp_stroke = StrokeOptions::lp_wide(Lp::points(2));
-        context.stroke_outline(style.outline, two_lp_stroke);
+        let outline_options = StrokeOptions::px_wide(
+            context
+                .get(&OutlineWidth)
+                .into_px(context.gfx.scale())
+                .ceil(),
+        );
+        context.stroke_outline(style.outline, outline_options);
 
         if context.focused(true) {
             if current_style == ButtonKind::Transparent {
-                let two_lp_stroke = two_lp_stroke.into_px(context.gfx.scale());
                 let focus_color = context.get(&HighlightColor);
                 // Some states of a transparent button have solid background
                 // colors. most_contrasting from a 0-alpha color is not a
@@ -389,17 +394,27 @@ impl Widget for Button {
                         .most_contrasting(&[focus_color, context.get(&TextColor)])
                 } else {
                     focus_color
+                }
+                .with_alpha(128);
+
+                let inset = (context.get(&IntrinsicPadding).into_px(context.gfx.scale())
+                    - outline_options.line_width)
+                    / 2;
+
+                let options = outline_options.colored(color);
+                let radii = context.get(&CornerRadius);
+                let radii = radii.map(|r| r.into_px(context.gfx.scale()));
+                let ring_rect =
+                    Rect::new(Point::squared(inset), context.gfx.region().size - inset * 2);
+
+                let focus_ring = if radii.is_zero() {
+                    Shape::stroked_rect(ring_rect, options.into_px(context.gfx.scale()))
+                } else {
+                    Shape::stroked_round_rect(ring_rect, radii, options)
                 };
-
-                let inset = context.get(&IntrinsicPadding).into_px(context.gfx.scale());
-
-                let focus_ring = Shape::stroked_rect(
-                    Rect::new(Point::squared(inset), context.gfx.region().size - inset * 2),
-                    two_lp_stroke.colored(color),
-                );
                 context.gfx.draw_shape(&focus_ring);
             } else if context.is_default() {
-                context.stroke_outline(context.get(&OutlineColor), two_lp_stroke);
+                context.stroke_outline(context.get(&OutlineColor), outline_options);
             } else {
                 context.draw_focus_ring();
             }